requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jq==1.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-text-splitters==0.3.6
langsmith==0.3.11
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.3
openai==1.65.1
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
patool==4.0.0
pgvector==0.3.6
pillow==11.1.0
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pypdf==5.3.0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-telegram-bot==21.10
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
youtube-transcript-api==0.6.3
zstandard==0.23.0
